+++ src/orchestrator/event/mod.rs
@@ -1,0 +1,3 @@
+// Expose only to crate
+pub(crate) mod click;
+pub(crate) mod keypress;
